Create /etc/mysql manually in debian/rules
We will use mysql-common.dirs, but as long as debian/rules manually
installs things into this directory before dh_installdirs is run, we
need to do it manually. Leaving this in temporarily so that the use of
mysql-common.dirs will work without breaking anything. When debian/rules
is cleaned up further and these install calls moved into .install files
then this can be removed at the same time.